requirements.txt
discord.py
discord
dotenv
pycryptodome